requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
dkimpy==1.1.8
dnspython==2.7.0
fastapi==0.115.4
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
pyspf==2.0.14
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
